+++ priv/repo/migrations/20230329025816_rename_type_to_class.exs
@@ -1,0 +1,7 @@
+defmodule Cannery.Repo.Migrations.RenameTypeToClass do
+  use Ecto.Migration
+
+  def change do
+    rename table(:ammo_types), :type, to: :class
+  end
+end